backend/requirements.txt
uvicorn[standard]==0.37
fastapi==0.116
openai==2.8.1
openai-chatkit==1.1.0
